APR default handler without circular deps
