+++ modules/pool/lib/apr-data-sources/yb-apr-handlers/sources/default-apr-handler.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { AprHandler } from '..';
+import type { AprHandler } from '../types';
 
 export class DefaultAprHandler implements AprHandler {
     tokenAddress: string;
